Fetching moves from PokeAPI

See merge request paul7098/project-pokemon-dawson!17

+++ app/src/main/java/ca/dawsoncollege/project_pokemon/Battle.kt
@@ -83,8 +83,8 @@ abstract class Battle(val playerTrainer: PlayerTrainer, val context: Context) {
     }
 
     // check if move succeeds
-    private fun moveSuccessCheck(accuracy: Double): Boolean {
-        val randNum = Random.nextDouble(0.0, 100.0)
+    private fun moveSuccessCheck(accuracy: Int): Boolean {
+        val randNum = Random.nextInt(0, 100)
         // check move success using probabilities
         if (accuracy >= randNum) {
             // move succeeds
